Create Chorizo umbrella project
Moves the ChorizoCore app into a subproject of the new umbrella project
