+++ apps/chorizo_core/config/config.exs
@@ -1,0 +1,34 @@
+# This file is responsible for configuring your application
+# and its dependencies with the aid of the Mix.Config module.
+use Mix.Config
+
+# This configuration is loaded before any dependency and is restricted
+# to this project. If another project depends on this project, this
+# file won't be loaded nor affect the parent project. For this reason,
+# if you want to provide default values for your application for
+# 3rd-party users, it should be done in your "mix.exs" file.
+
+# You can configure your application as:
+#
+#     config :chorizo_core, key: :value
+#
+# and access this configuration in your application as:
+#
+#     Application.get_env(:chorizo_core, :key)
+#
+# You can also configure a 3rd-party app:
+#
+#     config :logger, level: :info
+
+config :chorizo_core, ecto_repos: [ChorizoCore.Repositories.Repo]
+config :chorizo_core, ChorizoCore.Repositories.Repo,
+  migration_primary_key: [type: :binary_id],
+  migration_timestamps: [type: :utc_datetime]
+
+# It is also possible to import configuration files, relative to this
+# directory. For example, you can emulate configuration per environment
+# by uncommenting the line below and defining dev.exs, test.exs and such.
+# Configuration from the imported file will override the ones defined
+# here (which is why it is important to import them last).
+#
+import_config "#{Mix.env}.exs"
