+++ apps/chorizo_core/lib/chorizo_core/chore_management.ex
