+++ apps/chorizo_core/lib/chorizo_core/entities/chore.ex
