+++ apps/chorizo_core/lib/chorizo_core/entities/user.ex
