+++ apps/chorizo_core/lib/chorizo_core/entities/uuid.ex
