+++ apps/chorizo_core/scripts/update_master_documentation.sh
