+++ apps/chorizo_core/test/chorizo_core/entities/user_test.exs
